separated all and my components
