+++ src/components/myApplyances/MyApplyanceDetails.js
