+++ src/components/views/UserViews.js
@@ -1,8 +1,12 @@
 import { Outlet, Route, Routes } from "react-router-dom"
 import { ApplyanceForm } from "../applyances/ApplyanceForm"
 import { ApplyanceContainer } from "../applyances/ApplyanceContainer"
-import { MyApplyanceContainer } from "../applyances/MyApplyanceContainer"
 import { HomePage } from "../home/HomePage"
+import { ApplyanceDetails } from "../applyances/ApplyanceDetails"
+// import { MyApplyanceDetails } from "../applyances/MyApplyanceDetails"
+// import { MyApplyanceContainer } from "../applyances/MyApplyanceContainer"
+import { MyApplyanceDetails } from "../myApplyances/MyApplyanceDetails"
+import { MyApplyanceContainer } from "../myApplyances/MyApplyanceContainer"
 
 export const UserViews = () => {
 
@@ -22,11 +26,12 @@ export const UserViews = () => {
 
                 {/* All ApplYances Page */}
                 <Route path="all-applyances" element={<ApplyanceContainer />} />
+                <Route path="all-applyances/:applyanceId" element={<ApplyanceDetails />} />
 
 
                 {/* My ApplYances Page */}
                 <Route path="my-applyances" element={<MyApplyanceContainer />} />
-                <Route path="my-applyances/:myApplyanceId" element={<MyApplyanceContainer />} />
+                <Route path="my-applyances/:myApplyanceId" element={<MyApplyanceDetails />} />
 
                 {/* Add New ApplYance Page */}
                 <Route path="new-applyances" element={<ApplyanceForm />} />
